pytorch/requirements.txt
aiofiles==0.8.0
aiohttp==4.0.0a1
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
bentoml==1.0.0a4
cattrs==1.10.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
circus==0.17.1
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
contextlib2==21.6.0
deepmerge==1.0.4a0
Deprecated==1.2.13
filelock==3.6.0
fs==2.4.15
h11==0.13.0
idna==3.3
multidict==4.7.6
numpy==1.22.2
opentelemetry-api==1.9.0
opentelemetry-instrumentation==0.28b0
opentelemetry-instrumentation-aiohttp-client==0.28b0
opentelemetry-instrumentation-asgi==0.28b0
opentelemetry-sdk==1.9.0
opentelemetry-semantic-conventions==0.28b0
opentelemetry-util-http==0.28b0
packaging==21.3
pathspec==0.9.0
pep517==0.12.0
Pillow==9.0.1
pip-tools==6.5.1
prometheus-client==0.13.1
psutil==5.9.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-json-logger==2.0.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
rich==11.2.0
schema==0.7.5
simple-di==0.1.4
six==1.16.0
sniffio==1.2.0
starlette==0.18.0
timm==0.5.4
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.8
uvicorn==0.17.5
wrapt==1.13.3
yarl==1.7.2
